Rename nn => moments
