+++ pyoneer/moments/__init__.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from pyoneer.nn.nn_impl import (
+from pyoneer.moments.moments_impl import (
     range_moments,
     StaticMoments,
     StreamingMoments,
